@@ -4,8 +4,7 @@ version: 3.0.0
 homepage: https://github.com/schibsted/tweet_ui
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.12.13+hotfix.5"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -20,8 +19,6 @@ dependencies:
   collection: ^1.15.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
   lint: ^1.5.3
 
 flutter:
