@@ -1,7 +1,7 @@
 import 'package:url_launcher/url_launcher.dart';
 
 Future<void> openUrl(String url) async {
-  if (await canLaunch(url)) {
-    await launch(url);
+  if (await canLaunchUrl(Uri.parse(url))) {
+    await launchUrl(Uri.parse(url));
   }
 }
